requirements.txt
absl-py==0.7.1
app==0.0.1
astor==0.8.0
boto3==1.9.212
census==0.8.13
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cloudinary==1.16.0
cycler==0.10.0
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
gast==0.2.2
google-pasta==0.1.7
graphviz==0.11.1
grpcio==1.22.0
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
jellyfish==0.5.6
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mock==3.0.5
numpy==1.16.4
pandas==0.24.2
Pillow==6.1.0
protobuf==3.8.0
pydotplus==2.0.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.5
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.3
us==1.0.0
Werkzeug==0.15.4
wrapt==1.11.2
